+++ examples/resources/conjur_swa_node_group/import.sh
@@ -1,0 +1,1 @@
+terraform import conjur_swa_node_group.my_node_group prod.example.org/prod-servers/prod-nodes