+++ examples/resources/conjur_swa_server/import.sh
@@ -1,0 +1,1 @@
+terraform import conjur_swa_server.my_server prod.example.org/prod-servers/my-workload