+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- New `conjur_swa_trust_domain` resource for managing SWA Trust Domains.
+- New `conjur_swa_server_group` resource for managing SWA Server Groups within a Trust Domain.
+- New `conjur_swa_node_group` resource for managing SWA Node Groups within a Server Group.
+- New `conjur_swa_server` resource for managing SWA Servers (agents) within a Server Group.
+
 ## [0.8.4] - 2026-03-25
 
 ### Security